+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Bump itext7 to 7.2.3
 - Fix app crashing when scheduling overdue download worker
 - Handle nullable fields when generating overdue CSV from database
+- Add support for selected download and share in overdue screen
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 - [In Progress: 12 Jul 2022] Add support for selected download and share in overdue screen